While loop and list and append function.
In this code, we run the while loop and basic concept of it. Secondly, we have store string file using the append function in the list. We did some experiment for commenting and uncommenting with use of ctr+o+/.   For selecting line movement together, we have used short key Ctrl + shift+ upper/lower arrow.