While loop, case and for loop
In this Program, we try to take some cases as an input then add it into a list. Secondly, We print the list directly but to avoid the list we use for loop and get the exact message in showing option. If the users give exit as an input, program break the loops and closed.